Create README.md
📚 Welcome to my repository of Simple Exercises! Here, you will find a collection of straightforward coding exercises designed to enhance your programming skills. 💪🏼💻

These exercises cover a wide range of topics and are perfect for beginners or anyone looking to practice and reinforce their understanding of fundamental concepts. 🌟 Each exercise comes with a clear problem statement and a sample solution to guide you. 📝

Feel free to explore, contribute, and challenge yourself with these simple yet effective exercises. 🚀 Happy coding! 💻✨